Split rtab.c/h and ca.c/h
